requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
descope>=1.5.0
python-multipart==0.0.6
requests>=2.28.0
python-dotenv==1.0.0
PyJWT>=2.0.0